Improve MSBuild search
- It split a `VswhereVersionLocator` from `CommandLineToolVersionLocator`.
- Introduce a `MSBuildVersionLocator` and expose it through
`VisualStudioLocatorTestFixture`
