+++ subprojects/internal-testing/src/main/groovy/org/gradle/util/TestPrecondition.groovy
@@ -150,6 +150,10 @@ enum TestPrecondition implements org.gradle.internal.Factory<Boolean> {
     XCODE({
         // Simplistic approach at detecting Xcode by assuming macOS imply Xcode is present
         MAC_OS_X.fulfilled
+    }),
+    MSBUILD({
+        // Simplistic approach at detecting MSBuild by assuming Windows imply MSBuild is present
+        WINDOWS.fulfilled
     })
 
     /**
